tests: Move JSON schema validation tests to common location
Moves the tests for the JSON Schema validation to a common location so
that they can be used by multiple tests. This also gives the tests
useful descriptions so it is easier to identify which test has failed.
