+++ tests/test_jsonschema.py
@@ -9,6 +9,7 @@ import json
 import jsonschema
 import pytest
 from pathlib import Path
+import jsonvalidation
 
 THIS_FILE = Path(__file__)
 THIS_DIR = THIS_FILE.parent
@@ -132,613 +133,9 @@ class TestOutput:
             assert "\t" not in line, f"Line {num + 1} has tabs"
 
 
-CONTEXT = object()
-
-
-def replace_context(d, url):
-    for k, v in d.items():
-        if v is CONTEXT:
-            d[k] = url
-        elif isinstance(v, dict):
-            replace_context(v, url)
-
-
-BASE_OBJ = {
-    "@context": CONTEXT,
-    "@type": "test-class",
-}
-
-
-def node_kind_tests(name, blank, iri):
-    return [
-        (
-            blank,
-            {
-                "@context": CONTEXT,
-                "@type": name,
-            },
-        ),
-        (
-            blank,
-            {
-                "@context": CONTEXT,
-                "@type": name,
-                "@id": "_:blank",
-            },
-        ),
-        (
-            blank,
-            {
-                "@context": CONTEXT,
-                "@type": "link-class",
-                "link-class-link-prop": {
-                    "@type": name,
-                },
-            },
-        ),
-        (
-            blank,
-            {
-                "@context": CONTEXT,
-                "@type": "link-class",
-                "link-class-link-prop": {
-                    "@type": name,
-                    "@id": "_:blank",
-                },
-            },
-        ),
-        (
-            iri,
-            {
-                "@context": CONTEXT,
-                "@type": name,
-                "@id": "http://example.com/name",
-            },
-        ),
-        (
-            iri,
-            {
-                "@context": CONTEXT,
-                "@type": "link-class",
-                "link-class-link-prop": {
-                    "@type": name,
-                    "@id": "http://example.com/name",
-                },
-            },
-        ),
-    ]
-
-
-@pytest.mark.parametrize(
-    "passes,data",
-    [
-        # Empty graph
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@graph": [],
-            },
-        ),
-        # Missing context
-        (
-            False,
-            {
-                "@graph": [],
-            },
-        ),
-        # Bad context
-        (
-            False,
-            {
-                "@context": "http://foo.com",
-                "@graph": [],
-            },
-        ),
-        # Unknown root field
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@graph": [],
-                "foo": "bar",
-            },
-        ),
-        # Minimal with object
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "link-class",
-                    }
-                ],
-            },
-        ),
-        # Missing type
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@graph": [{}],
-            },
-        ),
-        # Unknown object field
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "link-class",
-                        "foo": "bar",
-                    }
-                ],
-            },
-        ),
-        # Nested base class
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "link-class",
-                        "link-class-link-prop": {
-                            "@type": "link-class",
-                        },
-                    }
-                ],
-            },
-        ),
-        # Nested derived class
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "link-class",
-                        "link-class-link-prop": {
-                            "@type": "link-derived-class",
-                        },
-                    }
-                ],
-            },
-        ),
-        # Derived class with nested base class
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "link-derived-class",
-                        "link-class-link-prop": {
-                            "@type": "link-class",
-                        },
-                    }
-                ],
-            },
-        ),
-        # Derived class with nested derived class
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "link-derived-class",
-                        "link-class-link-prop": {
-                            "@type": "link-derived-class",
-                        },
-                    }
-                ],
-            },
-        ),
-        # Link by string
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "link-class",
-                        "link-class-link-prop": "http://serialize.example.org/test",
-                    }
-                ],
-            },
-        ),
-        # Bad nested class type
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "link-class",
-                        "link-class-link-prop": {
-                            "@type": "test-another-class",
-                        },
-                    }
-                ],
-            },
-        ),
-        # Pattern test
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "test-class",
-                        "test-class/regex": "foo1",
-                        "test-class/regex-list": ["foo1", "foo2"],
-                    }
-                ],
-            },
-        ),
-        # Bad pattern
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "test-class",
-                        "test-class/regex": "fooa",
-                    }
-                ],
-            },
-        ),
-        # Bad pattern list
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "test-class",
-                        "test-class/regex-list": ["fooa"],
-                    }
-                ],
-            },
-        ),
-        # Empty list
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "test-class",
-                        "test-class/string-list-prop": [],
-                    }
-                ],
-            },
-        ),
-        # Required property
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "test-class-required",
-                        "test-class/required-string-scalar-prop": "hello",
-                        "test-class/required-string-list-prop": ["hello", "world"],
-                    }
-                ],
-            },
-        ),
-        # Missing required scalar
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "test-class-required",
-                        "test-class/required-string-list-prop": ["hello", "world"],
-                    }
-                ],
-            },
-        ),
-        # Missing required list
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "test-class-required",
-                        "test-class/required-string-scalar-prop": "hello",
-                    }
-                ],
-            },
-        ),
-        # Short list
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "test-class-required",
-                        "test-class/required-string-scalar-prop": "hello",
-                        "test-class/required-string-list-prop": [],
-                    }
-                ],
-            },
-        ),
-        # Long list
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@graph": [
-                    {
-                        "@type": "test-class-required",
-                        "test-class/required-string-scalar-prop": "hello",
-                        "test-class/required-string-list-prop": [
-                            "hello",
-                            "world",
-                            "too long",
-                        ],
-                    }
-                ],
-            },
-        ),
-        # Root object
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@type": "test-class",
-            },
-        ),
-        # Root object with unknown field
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@type": "test-class",
-                "foo": "bar",
-            },
-        ),
-        # Root object with missing context
-        (
-            False,
-            {
-                "@type": "test-class",
-            },
-        ),
-        # Root object with missing type
-        (
-            False,
-            {
-                "@context": CONTEXT,
-            },
-        ),
-        # Root object with bad context
-        (
-            False,
-            {
-                "@context": "http://foo.bar",
-                "@type": "test-class",
-            },
-        ),
-        # Root object with unknown type
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@type": "foo",
-            },
-        ),
-        # Node Kind tests
-        *node_kind_tests("node-kind-blank", True, False),
-        *node_kind_tests("node-kind-iri", False, True),
-        *node_kind_tests("node-kind-iri-or-blank", True, True),
-        *node_kind_tests("derived-node-kind-iri", False, True),
-        # Alternate ID
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@type": "id-prop-class",
-                "testid": "_:blank",
-            },
-        ),
-        # @id not allowed for alternate ID classes
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@type": "id-prop-class",
-                "@id": "_:blank",
-            },
-        ),
-        # Alternate ID in inherited class
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@type": "inherited-id-prop-class",
-                "testid": "_:blank",
-            },
-        ),
-        # @id not allowed for alternate ID classes
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@type": "inherited-id-prop-class",
-                "@id": "_:blank",
-            },
-        ),
-        # Extensible class
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@type": "extensible-class",
-                "extensible-class/required": "foo",
-                "link-class-link-prop": {
-                    "@type": "link-class",
-                },
-            },
-        ),
-        # Extensible class with custom type
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@type": "http://example.com/extended",
-                "extensible-class/required": "foo",
-                "link-class-link-prop": {
-                    "@type": "link-class",
-                },
-            },
-        ),
-        # Extended class with unknown property
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@type": "http://example.com/extended",
-                "extensible-class/required": "foo",
-                "unknown-prop": "foo",
-            },
-        ),
-        # Extended class with missing required property
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@type": "http://example.com/extended",
-                "unknown-prop": "foo",
-            },
-        ),
-        # Nested extensible class
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@type": "link-class",
-                "link-class-link-prop": {
-                    "@type": "extensible-class",
-                    "extensible-class/required": "foo",
-                },
-            },
-        ),
-        # Nested extensible class with custom type
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@type": "link-class",
-                "link-class-link-prop": {
-                    "@type": "http://example.com/extended",
-                    "extensible-class/required": "foo",
-                },
-            },
-        ),
-        # Nested extensible class with custom unknown property
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@type": "link-class",
-                "link-class-link-prop": {
-                    "@type": "http://example.com/extended",
-                    "extensible-class/required": "foo",
-                    "unknown-prop": "foo",
-                },
-            },
-        ),
-        # Nested extended class with missing required property
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@type": "link-class",
-                "link-class-link-prop": {
-                    "@type": "http://example.com/extended",
-                },
-            },
-        ),
-        # Abstract class
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@type": "abstract-class",
-            },
-        ),
-        # Concrete class derived from abstract class
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@type": "concrete-class",
-            },
-        ),
-        # Abstract SPDX style class
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@type": "abstract-spdx-class",
-            },
-        ),
-        # Concrete SPDX style class
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@type": "concrete-spdx-class",
-            },
-        ),
-        # An extensible abstract class cannot be instantiated
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@type": "extensible-abstract-class",
-            },
-        ),
-        # Any can type can be used where a extensible abstract class is
-        # references, except... (SEE NEXT)
-        (
-            True,
-            {
-                "@context": CONTEXT,
-                "@type": "uses-extensible-abstract-class",
-                "uses-extensible-abstract-class/prop": {
-                    "@type": "http://example.com/extended",
-                },
-            },
-        ),
-        # ... the exact type of the extensible abstract class is specifically
-        # not allowed
-        (
-            False,
-            {
-                "@context": CONTEXT,
-                "@type": "uses-extensible-abstract-class",
-                "uses-extensible-abstract-class/prop": {
-                    "@type": "extensible-abstract-class",
-                },
-            },
-        ),
-        # Base object for type tests
-        (True, BASE_OBJ),
-    ],
-)
+@jsonvalidation.validation_tests()
 def test_schema_validation(test_jsonschema, test_context_url, passes, data):
-    replace_context(data, test_context_url)
+    jsonvalidation.replace_context(data, test_context_url)
 
     if passes:
         jsonschema.validate(data, schema=test_jsonschema)
@@ -747,210 +144,9 @@ def test_schema_validation(test_jsonschema, test_context_url, passes, data):
             jsonschema.validate(data, schema=test_jsonschema)
 
 
-def type_tests(name, *, good=[], bad=[], typ=[]):
-    tests = [
-        # None is never allowed
-        (False, name, None),
-        # Empty object is not allowed for a scalar
-        (False, name, {}),
-    ]
-
-    def lst(v):
-        if list in typ:
-            return [v]
-        return v
-
-    is_list = list in typ
-    if is_list:
-        # No scalars allowed
-        tests.append((False, name, True))
-        tests.append((False, name, False))
-        tests.append((False, name, 1))
-        tests.append((False, name, 1.5))
-        tests.append((False, name, "foo"))
-
-        # Empty list is allowed
-        tests.append((True, name, []))
-    else:
-        # Lists not allowed
-        tests.append((False, name, []))
-
-    if bool not in typ:
-        tests.append((False, name, lst(True)))
-        tests.append((False, name, lst(False)))
-
-    if not (int in typ or float in typ):
-        tests.append((False, name, lst(1)))
-
-    if float not in typ:
-        tests.append((False, name, lst(1.5)))
-
-    if str not in typ:
-        tests.append((False, name, lst("foo")))
-
-    for g in good:
-        tests.append((True, name, g))
-
-    for b in bad:
-        tests.append((False, name, b))
-
-    return tests
-
-
-@pytest.mark.parametrize(
-    "passes,name,val",
-    [
-        *type_tests(
-            "test-class/nonnegative-integer-prop", good=[1, 0], bad=[-1], typ=[int]
-        ),
-        *type_tests("test-class/integer-prop", good=[1, 0, -1], typ=[int]),
-        *type_tests(
-            "test-class/float-prop",
-            good=[-1.5, 0, 1.5, "1.5", "0.0", "-1.5", "1", "-1"],
-            bad=["a1", "a", ""],
-            typ=[float],
-        ),
-        *type_tests(
-            "test-class/datetime-scalar-prop",
-            good=[
-                "2024-03-11T00:00:00+00:00",
-                "2024-03-11T00:00:00Z",
-                "2024-03-11T00:00:00",
-            ],
-            bad=["abc"],
-            typ=[str],
-        ),
-        *type_tests(
-            "test-class/datetimestamp-scalar-prop",
-            good=[
-                "2024-03-11T00:00:00+00:00",
-                "2024-03-11T00:00:00Z",
-            ],
-            bad=[
-                "abc",
-                "2024-03-11T00:00:00",
-            ],
-            typ=[str],
-        ),
-        *type_tests("test-class/string-scalar-prop", good=["foo", ""], typ=[str]),
-        *type_tests("test-class/named-property", good=["foo", ""], typ=[str]),
-        *type_tests(
-            "test-class/enum-prop",
-            good=["foo"],
-            bad=["enumType/foo"],
-            typ=[str],
-        ),
-        *type_tests(
-            "test-class/class-prop",
-            good=[
-                {"@type": "test-class"},
-                {"@type": "test-derived-class"},
-                "_:blanknode",
-                "http://serialize.example.org/test",
-                # Named individual
-                "named",
-            ],
-            bad=[
-                {"@type": "test-another-class"},
-                {"@type": "parent-class"},
-                "not/an/iri",
-                "not-an-iri",
-            ],
-            typ=[object, str],
-        ),
-        *type_tests(
-            "test-class/regex",
-            good=[
-                "foo1",
-                "foo2",
-                "foo2a",
-            ],
-            bad=[
-                "bar",
-                "fooa",
-                "afoo1",
-            ],
-            typ=[str],
-        ),
-        *type_tests(
-            "test-class/regex-datetime",
-            good=[
-                "2024-03-11T00:00:00+01:00",
-            ],
-            bad=[
-                "2024-03-11T00:00:00+1:00",
-                "2024-03-11T00:00:00+00:00",
-                "2024-03-11T00:00:00",
-                "2024-03-11T00:00:00Z",
-            ],
-            typ=[],
-        ),
-        *type_tests(
-            "test-class/regex-datetimestamp",
-            good=[
-                "2024-03-11T00:00:00Z",
-            ],
-            bad=[
-                "2024-03-11T00:00:00",
-                "2024-03-11T00:00:00+1:00",
-                "2024-03-11T00:00:00+00:00",
-            ],
-            typ=[],
-        ),
-        *type_tests(
-            "test-class/string-list-prop",
-            good=[["foo", "bar"], [""]],
-            typ=[str, list],
-        ),
-        *type_tests(
-            "test-class/datetime-list-prop",
-            good=[
-                ["2024-03-11T00:00:00Z"],
-            ],
-            bad=[
-                ["abc"],
-            ],
-            typ=[str, list],
-        ),
-        *type_tests(
-            "test-class/enum-list-prop",
-            good=[
-                [
-                    "foo",
-                    "bar",
-                    "nolabel",
-                ]
-            ],
-            bad=[
-                [
-                    "enumType/foo",
-                    "foo",
-                ]
-            ],
-            typ=[str, list],
-        ),
-        *type_tests(
-            "test-class/class-list-prop",
-            good=[
-                [
-                    {"@type": "test-class"},
-                    {"@type": "test-derived-class"},
-                    "_:blanknode",
-                    "http://serialize.example.org/test",
-                ]
-            ],
-            bad=[
-                [{"@type": "test-another-class"}],
-                [{"@type": "parent-class"}],
-            ],
-            typ=[object, str, list],
-        ),
-    ],
-)
-def test_schema_type_validation(test_jsonschema, test_context_url, passes, name, val):
-    data = BASE_OBJ.copy()
-    replace_context(data, test_context_url)
-    data[name] = val
+@jsonvalidation.type_tests()
+def test_schema_type_validation(test_jsonschema, test_context_url, passes, data):
+    jsonvalidation.replace_context(data, test_context_url)
 
     if passes:
         jsonschema.validate(data, schema=test_jsonschema)
